Refactored common menu items
