+++ apps/admin/src/layout/app-sidebar/AppSidebarFooter.tsx
@@ -6,8 +6,8 @@ import {
     SidebarMenu,
     SidebarMenuItem
 } from "@tryghost/shade"
-import UserMenu from "./UserMenu";
 import WhatsNewDialog from "@/whats-new/components/whats-new-dialog";
+import { UserMenu } from "./UserMenu";
 
 function AppSidebarFooter({ ...props }: React.ComponentProps<typeof SidebarFooter>) {
     const [isWhatsNewDialogOpen, setIsWhatsNewDialogOpen] = useState(false);
