+++ apps/admin/src/layout/app-sidebar/index.ts
@@ -1,2 +1,1 @@
 export {default} from './AppSidebar';
-export {default as ContributorProfileMenu} from './ContributorUserMenu';
